UnivagRH/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==7.1.2
click-didyoumean==0.3.1
colorama==0.4.6
distro==1.9.0
Flask==2.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
numpy==2.3.4
openai==2.6.1
p==1.5.0
pandas==2.3.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==2.0.3
gunicorn==21.2.0
python-dotenv==1.0.1
SQLAlchemy==1.4.54
Flask-SQLAlchemy==2.5.1
mysql-connector-python==9.5.0
Flask-Login==0.6.3
